chatbot_backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.0
fastapi-cli==0.0.5
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
mysql-connector==2.2.9
mysql-connector-python==9.0.0
openai==1.51.2
orjson==3.10.7
pydantic==2.9.2
pydantic-extra-types==2.9.0
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.2
redis==5.1.1
rich==13.9.2
shellingham==1.5.4
sniffio==1.3.1
starlette==0.38.6
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
ujson==5.10.0
uvicorn==0.31.0
watchfiles==0.24.0
websockets==13.1
